--- aufgabe3/SYSimulation/doc/allclasses-noframe.html
@@ -1,32 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!--NewPage-->
-<HTML>
-<HEAD>
-<!-- Generated by javadoc (build 1.6.0_22) on Tue Nov 01 18:37:59 CET 2011 -->
-<META http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<TITLE>
-All Classes
-</TITLE>
-
-<META NAME="date" CONTENT="2011-11-01">
-
-<LINK REL ="stylesheet" TYPE="text/css" HREF="stylesheet.css" TITLE="Style">
-
-
-</HEAD>
-
-<BODY BGCOLOR="white">
-<FONT size="+1" CLASS="FrameHeadingFont">
-<B>All Classes</B></FONT>
-<BR>
-
-<TABLE BORDER="0" WIDTH="100%" SUMMARY="">
-<TR>
-<TD NOWRAP><FONT CLASS="FrameItemFont"><A HREF="sim/SYSimulation.html" title="class in sim">SYSimulation</A>
-<BR>
-</FONT></TD>
-</TR>
-</TABLE>
-
-</BODY>
-</HTML>
