--- aufgabe3/SYSimulation/doc/package-list
@@ -1,1 +1,0 @@
-sim
