+++ aufgabe3/scotlandYard_2021/gui/MapEntry.java
