+++ aufgabe3/scotlandYard_2021/gui/stations.txt
